and what about that ?

+++ zds/tutorialv2/models/models_database.py
@@ -947,7 +947,9 @@ def delete_published_content_in_elasticsearch(sender, instance, **kwargs):
     print('d1')
 
     index_manager = ESIndexManager(**settings.ES_SEARCH_INDEX)
-    index_manager.delete_by_query(FakeChapter.get_es_document_type(), ES_Q('match', _routing=instance.es_id))
+
+    if index_manager.connected_to_es and index_manager.es.indices.exists(index_manager.index):
+        index_manager.delete_by_query(FakeChapter.get_es_document_type(), ES_Q('match', _routing=instance.es_id))
 
     print('d2')
